Moves core logic into its own package
This PR moves the core Include/Exclude logic into its own package. This allows the core logic to be imported and reused by other buildpacks or libraries. This core logic package does not depend on packit or libcnb/libpak so it can be used in either style buildpack.

Summary:

- Creates 'logic' package (not tied to that name)
- Moves exclude, include and tests into that package
- Updates buildpack & tests to use logic package
