+++ build.go
@@ -3,20 +3,22 @@ package sourceremoval
 import (
 	"os"
 
+	"github.com/paketo-buildpacks/source-removal/logic"
+
 	"github.com/paketo-buildpacks/packit/v2"
 )
 
 func Build() packit.BuildFunc {
 	return func(context packit.BuildContext) (packit.BuildResult, error) {
 		if includeVal, ok := os.LookupEnv("BP_INCLUDE_FILES"); ok {
-			err := Include(context.WorkingDir, includeVal)
+			err := logic.Include(context.WorkingDir, includeVal)
 			if err != nil {
 				return packit.BuildResult{}, err
 			}
 		}
 
 		if excludeVal, ok := os.LookupEnv("BP_EXCLUDE_FILES"); ok {
-			err := Exclude(context.WorkingDir, excludeVal)
+			err := logic.Exclude(context.WorkingDir, excludeVal)
 			if err != nil {
 				return packit.BuildResult{}, err
 			}
