+++ logic/exclude.go
@@ -1,4 +1,4 @@
-package sourceremoval
+package logic
 
 import (
 	"os"
