+++ logic/init_test.go
@@ -1,4 +1,4 @@
-package sourceremoval_test
+package logic_test
 
 import (
 	"testing"
@@ -9,7 +9,7 @@ import (
 
 func TestUnitSourceRemoval(t *testing.T) {
 	suite := spec.New("source-removal", spec.Report(report.Terminal{}))
-	suite("Build", testBuild)
-	suite("Detect", testDetect)
+	suite("Exclude", testExclude)
+	suite("Include", testInclude)
 	suite.Run(t)
 }
